fix: use u128 type for BPO update fraction constants

+++ crates/eips/src/eip7840.rs
@@ -83,7 +83,7 @@ impl BlobParams {
         Self {
             target_blob_count: eip7892::BPO1_TARGET_BLOBS_PER_BLOCK,
             max_blob_count: eip7892::BPO1_MAX_BLOBS_PER_BLOCK,
-            update_fraction: eip7892::BPO1_BASE_UPDATE_FRACTION as u128,
+            update_fraction: eip7892::BPO1_BASE_UPDATE_FRACTION,
             ..Self::osaka()
         }
     }
@@ -93,7 +93,7 @@ impl BlobParams {
         Self {
             target_blob_count: eip7892::BPO2_TARGET_BLOBS_PER_BLOCK,
             max_blob_count: eip7892::BPO2_MAX_BLOBS_PER_BLOCK,
-            update_fraction: eip7892::BPO2_BASE_UPDATE_FRACTION as u128,
+            update_fraction: eip7892::BPO2_BASE_UPDATE_FRACTION,
             ..Self::osaka()
         }
     }
