+++ crates/eips/src/eip7892.rs
@@ -10,7 +10,7 @@ pub const BPO1_TARGET_BLOBS_PER_BLOCK: u64 = 10;
 pub const BPO1_MAX_BLOBS_PER_BLOCK: u64 = 15;
 
 /// Update fraction for BPO1
-pub const BPO1_BASE_UPDATE_FRACTION: u64 = 8346193;
+pub const BPO1_BASE_UPDATE_FRACTION: u128 = 8346193;
 
 /// Targeted blob count with BPO2 activation
 pub const BPO2_TARGET_BLOBS_PER_BLOCK: u64 = 14;
@@ -19,7 +19,7 @@ pub const BPO2_TARGET_BLOBS_PER_BLOCK: u64 = 14;
 pub const BPO2_MAX_BLOBS_PER_BLOCK: u64 = 21;
 
 /// Update fraction for BPO2
-pub const BPO2_BASE_UPDATE_FRACTION: u64 = 11684671;
+pub const BPO2_BASE_UPDATE_FRACTION: u128 = 11684671;
 
 /// A scheduled blob parameter update entry.
 #[derive(Debug, Clone, PartialEq, Eq)]
